FEAT/BUILD: Add cmake option to disable X11
