+++ src/mumble/GlobalShortcut_x11.cpp
@@ -3,7 +3,7 @@
 // that can be found in the LICENSE file at the root of the
 // Mumble source tree or at <https://www.mumble.info/LICENSE>.
 
-#include "GlobalShortcut_unix.h"
+#include "GlobalShortcut_x11.h"
 
 #include "Settings.h"
 
@@ -16,11 +16,6 @@
 #ifndef NO_XINPUT2
 #	include <X11/extensions/XI2.h>
 #	include <X11/extensions/XInput2.h>
-#endif
-
-#ifdef Q_OS_LINUX
-#	include <fcntl.h>
-#	include <linux/input.h>
 #endif
 
 // We define a global macro called 'g'. This can lead to issues when included code uses 'g' as a type or parameter name
@@ -46,6 +41,7 @@
  * Returns a platform specific GlobalShortcutEngine object.
  *
  * @see GlobalShortcutX
+ * @see GlobalShortcutLinux
  * @see GlobalShortcutMac
  * @see GlobalShortcutWin
  */
@@ -63,27 +59,6 @@ GlobalShortcutX::GlobalShortcutX() {
 		qWarning("GlobalShortcutX: Unable to open dedicated display connection.");
 		return;
 	}
-
-#ifdef Q_OS_LINUX
-	if (g.s.bEnableEvdev) {
-		QString dir             = QLatin1String("/dev/input");
-		QFileSystemWatcher *fsw = new QFileSystemWatcher(QStringList(dir), this);
-		connect(fsw, SIGNAL(directoryChanged(const QString &)), this, SLOT(directoryChanged(const QString &)));
-		directoryChanged(dir);
-
-		if (qsKeyboards.isEmpty()) {
-			foreach (QFile *f, qmInputDevices)
-				delete f;
-			qmInputDevices.clear();
-
-			delete fsw;
-			qWarning(
-				"GlobalShortcutX: Unable to open any keyboard input devices under /dev/input, falling back to XInput");
-		} else {
-			return;
-		}
-	}
-#endif
 
 	for (int i = 0; i < ScreenCount(display); ++i)
 		qsRootWindows.insert(RootWindow(display, i));
@@ -245,111 +220,6 @@ void GlobalShortcutX::displayReadyRead(int) {
 
 		XFreeEventData(display, cookie);
 	}
-#endif
-}
-
-// One of the raw /dev/input devices has ready input
-void GlobalShortcutX::inputReadyRead(int) {
-#ifdef Q_OS_LINUX
-	if (!g.s.bEnableEvdev) {
-		return;
-	}
-
-	struct input_event ev;
-
-	if (bNeedRemap)
-		remap();
-
-	QFile *f = qobject_cast< QFile * >(sender()->parent());
-	if (!f)
-		return;
-
-	bool found = false;
-
-	while (f->read(reinterpret_cast< char * >(&ev), sizeof(ev)) == sizeof(ev)) {
-		found = true;
-		if (ev.type != EV_KEY)
-			continue;
-		bool down;
-		switch (ev.value) {
-			case 0:
-				down = false;
-				break;
-			case 1:
-				down = true;
-				break;
-			default:
-				continue;
-		}
-		int evtcode = ev.code + 8;
-		handleButton(evtcode, down);
-	}
-
-	if (!found) {
-		int fd      = f->handle();
-		int version = 0;
-		if ((ioctl(fd, EVIOCGVERSION, &version) < 0) || (((version >> 16) & 0xFF) < 1)) {
-			qWarning("GlobalShortcutX: Removing dead input device %s", qPrintable(f->fileName()));
-			qmInputDevices.remove(f->fileName());
-			qsKeyboards.remove(f->fileName());
-			delete f;
-		}
-	}
-#endif
-}
-
-#define test_bit(bit, array) (array[bit / 8] & (1 << (bit % 8)))
-
-// The /dev/input directory changed
-void GlobalShortcutX::directoryChanged(const QString &dir) {
-#ifdef Q_OS_LINUX
-	if (!g.s.bEnableEvdev) {
-		return;
-	}
-
-	QDir d(dir, QLatin1String("event*"), 0, QDir::System);
-	foreach (QFileInfo fi, d.entryInfoList()) {
-		QString path = fi.absoluteFilePath();
-		if (!qmInputDevices.contains(path)) {
-			QFile *f = new QFile(path, this);
-			if (f->open(QIODevice::ReadOnly)) {
-				int fd = f->handle();
-				int version;
-				char name[256];
-				uint8_t events[EV_MAX / 8 + 1];
-				memset(events, 0, sizeof(events));
-				if ((ioctl(fd, EVIOCGVERSION, &version) >= 0) && (ioctl(fd, EVIOCGNAME(sizeof(name)), name) >= 0)
-					&& (ioctl(fd, EVIOCGBIT(0, sizeof(events)), &events) >= 0) && test_bit(EV_KEY, events)
-					&& (((version >> 16) & 0xFF) > 0)) {
-					name[255] = 0;
-					qWarning("GlobalShortcutX: %s: %s", qPrintable(f->fileName()), name);
-					// Is it grabbed by someone else?
-					if ((ioctl(fd, EVIOCGRAB, 1) < 0)) {
-						qWarning("GlobalShortcutX: Device exclusively grabbed by someone else (X11 using "
-								 "exclusive-mode evdev?)");
-						delete f;
-					} else {
-						ioctl(fd, EVIOCGRAB, 0);
-						uint8_t keys[KEY_MAX / 8 + 1];
-						if ((ioctl(fd, EVIOCGBIT(EV_KEY, sizeof(keys)), &keys) >= 0) && test_bit(KEY_SPACE, keys))
-							qsKeyboards.insert(f->fileName());
-
-						fcntl(f->handle(), F_SETFL, O_NONBLOCK);
-						connect(new QSocketNotifier(f->handle(), QSocketNotifier::Read, f), SIGNAL(activated(int)),
-								this, SLOT(inputReadyRead(int)));
-
-						qmInputDevices.insert(f->fileName(), f);
-					}
-				} else {
-					delete f;
-				}
-			} else {
-				delete f;
-			}
-		}
-	}
-#else
-	Q_UNUSED(dir);
 #endif
 }
